pyproject.toml
[project]
name = "sandbox-wandb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grad-cam>=1.5.5",
    "hydra-core>=1.3.2",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "mlflow>=3.4.0",
    "nbformat>=5.10.4",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "wandb>=0.21.3",
]
